fold vec_eval_scheduler back into the eval code; drop standalone module
Per the end-of-project plan, inline the extracted orchestration logic back where it is used
instead of leaving a standalone utils module:
- evaluate_instances_batched -> module function in eval/adapter.py (its only caller is run()).
- compute_q_score -> module function in metrics/task_metric.py (its only caller is TaskMetric;
  also avoids a circular import with adapter). Matches where the pre-refactor formula lived.
- delete OmniGibson/omnigibson/eval/utils/vec_eval_scheduler.py and its unit test
  tests/test_vec_eval_scheduler.py; repoint test_vec_eval_qscore_n2.py's import to task_metric;
  fix the stale module reference in behavior_task.get_goal_option_satisfaction docstring.

Verified: ruff clean, byte-compile, and an import+logic sanity check (compute_q_score cases +
evaluate_instances_batched batch/barrier/freeze) pass.
