+++ OmniGibson/omnigibson/tasks/behavior_task.py
@@ -460,7 +460,7 @@ class BehaviorTask(BaseTask):
         Per-env, per-goal-option predicate satisfaction, evaluated against env @env_idx's own object
         scope. This is the env-aware building block for the partial-success (Q-score) metric in
         vectorized evaluation: each goal-state option is evaluated independently so partial credit can
-        be computed (see ``omnigibson.eval.utils.vec_eval_scheduler.compute_q_score`` and ``TaskMetric``).
+        be computed (see ``omnigibson.metrics.task_metric.compute_q_score`` and ``TaskMetric``).
 
         Unlike reading ``ground_goal_state_options[*].evaluate()`` directly (which binds the single,
         shared ``compiled_task`` scope and therefore returns env 0's result for every env), this routes
